+++ .github/workflows/test.yml
@@ -20,6 +20,6 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v3
       - name: Smoke test
-        run: ./gradlew :run --args="listx"
+        run: ./gradlew :run --args="list"
       - name: Lint
         run: ./gradlew :detekt
