+++ build.gradle.kts
@@ -16,6 +16,7 @@ repositories {
 dependencies {
     implementation("com.github.ajalt.clikt:clikt:4.2.1")
     testImplementation("org.jetbrains.kotlin:kotlin-test")
+    detektPlugins("io.gitlab.arturbosch.detekt:detekt-formatting:1.23.3")
 }
 
 tasks.test {
