+++ src/main/kotlin/commands/List.kt
@@ -1,0 +1,9 @@
+package twinklecli.commands
+
+import com.github.ajalt.clikt.core.CliktCommand
+
+class List : CliktCommand(help = "List all apps") {
+    override fun run() {
+        echo("listing all apps")
+    }
+}
