+++ src/main/kotlin/commands/TwinkleCLI.kt
@@ -1,0 +1,7 @@
+package twinklecli.commands
+
+import com.github.ajalt.clikt.core.CliktCommand
+
+class TwinkleCLI : CliktCommand() {
+    override fun run() = Unit
+}
